Plan 01: targeted API fixes, consistency sweep, build pinning
Implements plans/01-part1-targeted-fixes.md (plan docs included):
- Upload/ShaderBinaryInfo take span<const u8>
- PipelineShaderStageInfo::Module is const Shader& (was raw Shader*)
- DescriptorSet asserts on sparse bindings and unsupported descriptor types
- WindowInfo::Extent is uvec2
- GetName() returns const string& everywhere; ImGUI -> ImGui renames
- Explicit ePrimary default; const format getters fed from ContextInfo
- SwapChain invalidation callbacks are a vector (AddInvalidationCallback)
- FetchContent deps pinned (VMA v3.4.0, nfd v1.3.0, stb/imnodes SHAs);
  vestigial /opt/homebrew/lib link directory removed
- Image::Download sizes its staging buffer by format block size
  (was hardcoded 4 bytes/pixel, broken for RGBA16F)

+++ include/Veng/Window.h
@@ -19,7 +19,7 @@ namespace Veng
 
     struct WindowInfo
     {
-        vec2 Extent;
+        uvec2 Extent;
         bool Resizable;
         EventCallback EventCallback;
         string Title;
